Also check $domain

+++ lets-encrypt/activate-ssl.sh
@@ -114,25 +114,32 @@ fi
 
 if [ $(nmap -sS -p 443 "$WANIP4" | grep -c "open") -eq 1 ]
 then
-  echo -e "\e[32mPort 443 is open!\e[0m"
+  echo -e "\e[32mPort 443 is open on $WANIP4!\e[0m"
   apt remove --purge nmap -y
 else
-  echo "Port 443 is not open. Please follow this guide to open ports in your router: https://www.techandme.se/open-port-80-443/"
+  echo "Port 443 is not open on $WANIP4. Please follow this guide to open ports in your router: https://www.techandme.se/open-port-80-443/"
   echo -e "\e[32m"
-  read -p "Press any key to continue... " -n1 -s
+  read -p "Press any key to test $domain... " -n1 -s
+  echo -e "\e[0m"
+  if [[ $(nc -z $domain 443) -eq 0 ]]
+  then
+  echo "Port 443 is not open on $domain. Please follow this guide to open ports in your router: https://www.techandme.se/open-port-80-443/"
+  echo -e "\e[32m"
+  read -p "Press any key to exit... " -n1 -s
   echo -e "\e[0m"
   apt remove --purge nmap -y
   exit 1
+  fi
 fi
 
 # Fetch latest version of test-new-config.sh
 if [ -f $SCRIPTS/test-new-config.sh ]
 then
     rm $SCRIPTS/test-new-config.sh
-    wget https://raw.githubusercontent.com/nextcloud/vm/master/lets-encrypt/test-new-config.sh -P $SCRIPTS
+    wget -q https://raw.githubusercontent.com/nextcloud/vm/master/lets-encrypt/test-new-config.sh -P $SCRIPTS
     chmod +x $SCRIPTS/test-new-config.sh
 else
-    wget https://raw.githubusercontent.com/nextcloud/vm/master/lets-encrypt/test-new-config.sh -P $SCRIPTS
+    wget -q https://raw.githubusercontent.com/nextcloud/vm/master/lets-encrypt/test-new-config.sh -P $SCRIPTS
     chmod +x $SCRIPTS/test-new-config.sh
 fi
